+++ orm/src/main/java/org/hibernate/tool/hbmlint/detector/BadCachingDetector.java
@@ -5,7 +5,7 @@ import org.hibernate.mapping.PersistentClass;
 import org.hibernate.mapping.Property;
 import org.hibernate.mapping.Value;
 import org.hibernate.tool.hbm2x.visitor.EntityNameFromValueVisitor;
-import org.hibernate.tool.hbmlint.Issue;
+import org.hibernate.tool.internal.export.lint.Issue;
 import org.hibernate.tool.internal.export.lint.IssueCollector;
 
 public class BadCachingDetector extends EntityModelDetector {
