+++ orm/src/main/java/org/hibernate/tool/hbmlint/detector/InstrumentationDetector.java
@@ -7,7 +7,7 @@ import org.hibernate.cfg.Environment;
 import org.hibernate.engine.spi.Managed;
 import org.hibernate.mapping.PersistentClass;
 import org.hibernate.mapping.Property;
-import org.hibernate.tool.hbmlint.Issue;
+import org.hibernate.tool.internal.export.lint.Issue;
 import org.hibernate.tool.internal.export.lint.IssueCollector;
 
 public class InstrumentationDetector extends EntityModelDetector {
