+++ orm/src/main/java/org/hibernate/tool/internal/export/lint/IssueCollector.java
@@ -1,7 +1,5 @@
 package org.hibernate.tool.internal.export.lint;
 
-import org.hibernate.tool.hbmlint.Issue;
-
 public interface IssueCollector {
 
 	public abstract void reportIssue(Issue analyze);
