+++ test/common/src/main/java/org/hibernate/tool/hbmlint/SchemaAnalyzer/TestCase.java
@@ -11,8 +11,8 @@ import java.util.List;
 import org.hibernate.boot.Metadata;
 import org.hibernate.boot.MetadataSources;
 import org.hibernate.mapping.Table;
-import org.hibernate.tool.hbmlint.Issue;
 import org.hibernate.tool.hbmlint.detector.SchemaByMetaDataDetector;
+import org.hibernate.tool.internal.export.lint.Issue;
 import org.hibernate.tool.internal.export.lint.IssueCollector;
 import org.hibernate.tools.test.util.JdbcUtil;
 import org.junit.After;
